chat_to_pdf/requirements.txt
aiofiles
altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
ci-info
click
colorama
configobj
configparser
distro
etelemetry
filelock
frontend
gitdb
GitPython
groq
h11
httpcore
httplib2
httpx
idna
isodate
itsdangerous
Jinja2
json_repair
jsonschema
jsonschema-specifications
looseversion
lxml
markdown-it-py
MarkupSafe
mdurl
narwhals
networkx
nibabel
nipype
numpy
packaging
pandas
pathlib
pillow
protobuf
prov
pyarrow
pydantic
pydantic_core
pydeck
pydot
Pygments
PyMuPDF
pyparsing
python-dateutil
python-dotenv
pytz
pyxnat
rdflib
referencing
requests
rich
rpds-py
scipy
simplejson
six
smmap
sniffio
starlette
streamlit
tenacity
toml
tornado
typing_extensions
tzdata
urllib3
uv
uvicorn
watchdog